Switched to promise model rather than callback model
 * Moved model to /lib/model.js
 * Created promise wrapper at index.js
 * Updated tests to reflect promise model
